… egret/model_library/unit_commitment/power_balance.py
… egret/models/tests/test_dcopf.py
… egret/models/tests/test_unit_commitment.py
… egret/models/unit_commitment.py