@bknueven 2020-11-19
Should switch to using `scipy.sparse.linalg.SuperLU` everywhere, which has algorithmic implications.